задатки JWT токенов, пока все работает

+++ src/main/java/com/team8/fooddelivery/service/CartService.java
@@ -20,7 +20,7 @@ public interface CartService {
 
     Cart clear(Long clientId);
 
-    double calculateTotal(Long clientId);
+    Long calculateTotal(Long clientId);
 }
